Shift recursive types to types.py
